Colonists now perform their assigned job to completion. Some jobs don't work yet, and skills are not updated after completion.

+++ Assets/Code/Job/Jobs/IJobDefinition.cs
@@ -14,7 +14,7 @@ namespace Snowship.NJob
 		// Job Definition Properties
 		Func<TileManager.Tile, int, bool>[] SelectionConditions { get; }
 		List<ResourceAmount> BaseRequiredResources { get; }
-		float TimeToWork { get; }
+		int TimeToWork { get; }
 		bool Returnable { get; }
 		int Layer { get; }
 		SelectionType SelectionType { get; }
